cv_bridge: Wrap divisions in int() for tests

+++ cv_bridge/test/python_bindings.py
@@ -14,11 +14,11 @@ def test_cvtColorForDisplay():
     label_value = 0
     grid_num_y, grid_num_x = 3, 4
     for grid_row in xrange(grid_num_y):
-        grid_size_y = height / grid_num_y
+        grid_size_y = int(height / grid_num_y)
         min_y = grid_size_y * grid_row
         max_y = min_y + grid_size_y
         for grid_col in xrange(grid_num_x):
-            grid_size_x = width / grid_num_x
+            grid_size_x = int(width / grid_num_x)
             min_x = grid_size_x * grid_col
             max_x = min_x + grid_size_x
             label[min_y:max_y, min_x:max_x] = label_value
